WIP Revamp serialization to better represent itself be better managed. Signed-off-by: Narimm <benjicharlton@gmail.com>
