+++ src/main/java/me/botsko/prism/actions/ItemStackAction.java
@@ -5,6 +5,8 @@ import me.botsko.prism.actionlibs.QueryParameters;
 import me.botsko.prism.appliers.ChangeResult;
 import me.botsko.prism.appliers.ChangeResultType;
 import me.botsko.prism.appliers.PrismProcessType;
+import me.botsko.prism.serializers.SerializationHandler;
+import me.botsko.prism.serializers.items.ItemStackSerializer;
 import me.botsko.prism.utils.EntityUtils;
 import me.botsko.prism.utils.InventoryUtils;
 import me.botsko.prism.utils.ItemUtils;
@@ -39,6 +41,7 @@ import org.bukkit.inventory.meta.FireworkEffectMeta;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.LeatherArmorMeta;
 import org.bukkit.inventory.meta.SkullMeta;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -46,18 +49,27 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Objects;
 
 public class ItemStackAction extends GenericAction {
 
     protected ItemStack item;
-    private ItemStackActionData actionData;
-    private Map<Enchantment, Integer> enchantments;
+    private ItemStackSerializer actionData;
+
     /**
      * Holds durability if no actionData yet exists.
      */
     private short tempDurability = -1;
+
+    private static ItemStackSerializer createItemStackActionData(ItemStack item, @Nullable  Map<Enchantment, Integer> enchantments) {
+        ItemStackSerializer data = ItemStackSerializer.createItemStackSerialized(item);
+        //overwrite actionData with provided enchants.
+        if (enchantments != null && !enchantments.isEmpty()) {
+            List<String> out = new ArrayList<>();
+            enchantments.forEach((enchantment, integer) -> out.add(enchantment.getKey().getKey() + ":" + integer));
+            data.enchs = out.toArray(new String[0]);
+        }
+        return data;
+    }
 
     @Override
     public boolean hasExtraData() {
@@ -83,161 +95,26 @@ public class ItemStackAction extends GenericAction {
 
     /**
      * Set the item.
-     * @param item ItemStack
-     * @param quantity int
+     *
+     * @param item         ItemStack
+     * @param quantity     int
      * @param enchantments Map of enchants.
      */
     public void setItem(ItemStack item, int quantity, Map<Enchantment, Integer> enchantments) {
-
-        actionData = new ItemStackActionData();
-
-        if (enchantments != null) {
-            this.enchantments = enchantments;
-        }
-
         if (item == null || item.getAmount() <= 0) {
             this.setCanceled(true);
             return;
         }
-
         this.item = item;
-        if (enchantments == null) {
-            this.enchantments = item.getEnchantments();
-        }
-
+        actionData = createItemStackActionData(item,enchantments);
         // Set basics
         setMaterial(item.getType());
-        actionData.durability = (short) ItemUtils.getItemDamage(item);
-
+        //override durability if a temp value was set.
         if (tempDurability >= 0) {
             actionData.durability = tempDurability;
             tempDurability = -1;
         }
-
         actionData.amt = quantity;
-
-        final ItemMeta meta = item.hasItemMeta() ? item.getItemMeta() : null;
-        if (meta != null) {
-            actionData.name = meta.getDisplayName();
-        }
-        if (meta instanceof LeatherArmorMeta) {
-            final LeatherArmorMeta lam = (LeatherArmorMeta) meta;
-            actionData.color = lam.getColor().asRGB();
-        } else if (meta instanceof SkullMeta) {
-            final SkullMeta skull = (SkullMeta) meta;
-            if (skull.hasOwner()) {
-                actionData.owner = Objects.requireNonNull(skull.getOwningPlayer()).getUniqueId().toString();
-            }
-        }
-
-        // Written books
-        if (meta instanceof BookMeta) {
-            final BookMeta bookMeta = (BookMeta) meta;
-            actionData.by = bookMeta.getAuthor();
-            actionData.title = bookMeta.getTitle();
-            actionData.content = bookMeta.getPages().toArray(new String[0]);
-        }
-
-        // Lore
-        if (meta != null && meta.hasLore()) {
-            actionData.lore = Objects.requireNonNull(meta.getLore()).toArray(new String[0]);
-        }
-
-        // Enchantments
-        if (!this.enchantments.isEmpty()) {
-            final String[] enchs = new String[this.enchantments.size()];
-            int i = 0;
-            for (final Entry<Enchantment, Integer> ench : this.enchantments.entrySet()) {
-                // This is silly
-                enchs[i] = ench.getKey().getKey().getKey() + ":" + ench.getValue();
-                i++;
-            }
-            actionData.enchs = enchs;
-        } else if (meta instanceof EnchantmentStorageMeta) {
-            final EnchantmentStorageMeta bookEnchantments = (EnchantmentStorageMeta) meta;
-            if (bookEnchantments.hasStoredEnchants()) {
-                if (bookEnchantments.getStoredEnchants().size() > 0) {
-                    final String[] enchs = new String[bookEnchantments.getStoredEnchants().size()];
-                    int i = 0;
-                    for (final Entry<Enchantment, Integer> ench : bookEnchantments.getStoredEnchants().entrySet()) {
-                        // This is absolutely silly
-                        enchs[i] = ench.getKey().getKey().getKey() + ":" + ench.getValue();
-                        i++;
-                    }
-                    actionData.enchs = enchs;
-                }
-            }
-        }
-        if (meta instanceof FireworkEffectMeta) {
-            applyFireWorksMetaToActionData(meta);
-        }
-        if (meta instanceof BannerMeta) {
-            List<Pattern> patterns = ((BannerMeta) meta).getPatterns();
-            Map<String, String> stringyPatterns = new HashMap<>();
-            patterns.forEach(
-                  pattern -> stringyPatterns.put(pattern.getPattern().getIdentifier(), pattern.getColor().name()));
-            actionData.bannerMeta = stringyPatterns;
-        }
-    }
-
-    private void applyFireWorksMetaToActionData(ItemMeta meta) {
-        final FireworkEffectMeta fireworkMeta = (FireworkEffectMeta) meta;
-        if (fireworkMeta.hasEffect()) {
-            final FireworkEffect effect = fireworkMeta.getEffect();
-            if (effect != null) {
-                if (!effect.getColors().isEmpty()) {
-                    final int[] effectColors = new int[effect.getColors().size()];
-                    int i = 0;
-                    for (final Color effectColor : effect.getColors()) {
-                        effectColors[i] = effectColor.asRGB();
-                        i++;
-                    }
-                    actionData.effectColors = effectColors;
-                }
-
-                if (!effect.getFadeColors().isEmpty()) {
-                    final int[] fadeColors = new int[effect.getColors().size()];
-                    final int i = 0;
-                    for (final Color fadeColor : effect.getFadeColors()) {
-                        fadeColors[i] = fadeColor.asRGB();
-                    }
-                    actionData.fadeColors = fadeColors;
-                }
-                if (effect.hasFlicker()) {
-                    actionData.hasFlicker = true;
-                }
-                if (effect.hasTrail()) {
-                    actionData.hasTrail = true;
-                }
-            }
-        }
-    }
-
-    private static ItemStack deserializeFireWorksMeta(ItemStack item, ItemMeta meta, ItemStackActionData actionData) {
-
-        final FireworkEffectMeta fireworkMeta = (FireworkEffectMeta) meta;
-        final Builder effect = FireworkEffect.builder();
-
-        for (int i = 0; i < actionData.effectColors.length; i++) {
-            effect.withColor(Color.fromRGB(actionData.effectColors[i]));
-        }
-        fireworkMeta.setEffect(effect.build());
-
-        if (actionData.fadeColors != null) {
-            for (int i = 0; i < actionData.fadeColors.length; i++) {
-                effect.withFade(Color.fromRGB(actionData.fadeColors[i]));
-            }
-            fireworkMeta.setEffect(effect.build());
-        }
-        if (actionData.hasFlicker) {
-            effect.flicker(true);
-        }
-        if (actionData.hasTrail) {
-            effect.trail(true);
-        }
-        fireworkMeta.setEffect(effect.build());
-        item.setItemMeta(fireworkMeta);
-        return item;
     }
 
     public void setSlot(String slot) {
@@ -246,7 +123,7 @@ public class ItemStackAction extends GenericAction {
 
     @Override
     public String serialize() {
-        return gson().toJson(actionData);
+        return SerializationHandler.gson().toJson(actionData);
     }
 
     @Override
@@ -254,99 +131,18 @@ public class ItemStackAction extends GenericAction {
         if (data == null || !data.startsWith("{")) {
             return;
         }
-        actionData = gson().fromJson(data, ItemStackActionData.class);
+        actionData = ItemStackSerializer.deserialize(data);
+        item = actionData.toBukkit();
 
-        item = new ItemStack(getMaterial(), actionData.amt);
-
-        ItemUtils.setItemDamage(item, actionData.durability);
-
-        // Restore enchantment
-        if (actionData.enchs != null && actionData.enchs.length > 0) {
-            for (final String ench : actionData.enchs) {
-                final String[] enchArgs = ench.split(":");
-                Enchantment enchantment = Enchantment.getByKey(NamespacedKey.minecraft(enchArgs[0]));
-
-                // Restore book enchantment
-                if (enchantment != null) {
-                    if (item.getType() == Material.ENCHANTED_BOOK) {
-                        final EnchantmentStorageMeta bookEnchantments = (EnchantmentStorageMeta) item.getItemMeta();
-                        bookEnchantments.addStoredEnchant(enchantment, Integer.parseInt(enchArgs[1]), false);
-                        item.setItemMeta(bookEnchantments);
-                    } else {
-                        item.addUnsafeEnchantment(enchantment, Integer.parseInt(enchArgs[1]));
-                    }
-                }
-            }
-        }
-
-        ItemMeta meta = item.getItemMeta();
-
-        // Leather color
-        if (meta instanceof LeatherArmorMeta && actionData.color > 0) {
-            final LeatherArmorMeta lam = (LeatherArmorMeta) meta;
-            lam.setColor(Color.fromRGB(actionData.color));
-            item.setItemMeta(lam);
-        } else if (meta instanceof SkullMeta && actionData.owner != null) {
-            final SkullMeta skull = (SkullMeta) meta;
-            skull.setOwningPlayer(Bukkit.getOfflinePlayer(EntityUtils.uuidOf(actionData.owner)));
-            item.setItemMeta(skull);
-        } else if (meta instanceof BookMeta) {
-            final BookMeta bookMeta = (BookMeta) meta;
-            bookMeta.setAuthor(actionData.by);
-            bookMeta.setTitle(actionData.title);
-            bookMeta.setPages(actionData.content);
-            item.setItemMeta(bookMeta);
-        }
-        if (meta instanceof FireworkEffectMeta && actionData.effectColors != null
-                && actionData.effectColors.length > 0) {
-
-            item = deserializeFireWorksMeta(item, meta, actionData);
-        }
-        if (meta instanceof BannerMeta && actionData.bannerMeta != null) {
-            Map<String, String> stringStringMap = actionData.bannerMeta;
-            List<Pattern> patterns = new ArrayList<>();
-            stringStringMap.forEach((patternIdentifier, dyeName) -> {
-                PatternType type = PatternType.getByIdentifier(patternIdentifier);
-                DyeColor color = DyeColor.valueOf(dyeName);
-                if (type != null && color != null) {
-                    Pattern p = new Pattern(color, type);
-                    patterns.add(p);
-                }
-            });
-            ((BannerMeta) meta).setPatterns(patterns);
-        }
-
-        if (actionData.name != null) {
-            if (meta == null) {
-                meta = item.getItemMeta();
-            }
-
-            if (meta != null) {
-                meta.setDisplayName(actionData.name);
-            }
-        }
-
-        if (actionData.lore != null) {
-            if (meta == null) {
-                meta = item.getItemMeta();
-            }
-
-            if (meta != null) {
-                meta.setLore(Arrays.asList(actionData.lore));
-            }
-        }
-
-        if (meta != null) {
-            item.setItemMeta(meta);
-        }
     }
 
-    public ItemStackActionData getActionData() {
+    public ItemStackSerializer getActionData() {
         return this.actionData;
     }
 
     /**
      * ItemStack.
+     *
      * @return ItemStack
      */
     public ItemStack getItem() {
@@ -355,6 +151,7 @@ public class ItemStackAction extends GenericAction {
 
     /**
      * Nice name.
+     *
      * @return String
      */
     @Override
@@ -630,24 +427,4 @@ public class ItemStackAction extends GenericAction {
         return new ChangeResult(result, null);
     }
 
-    @SuppressWarnings("WeakerAccess")
-    public static class ItemStackActionData {
-        public int amt;
-        public String name;
-        public int color;
-        public String owner;
-        public String[] enchs;
-        public String by;
-        public String title;
-        public String[] lore;
-        public String[] content;
-        public String slot = "-1";
-        public int[] effectColors;
-        public int[] fadeColors;
-        public boolean hasFlicker;
-        public boolean hasTrail;
-        public short durability = 0;
-        public Map<String, String> bannerMeta;
-
-    }
 }
