+++ src/main/java/me/botsko/prism/serializers/entity/WolfSerializer.java
@@ -1,0 +1,27 @@
+package me.botsko.prism.serializers.entity;
+
+import me.botsko.prism.utils.MiscUtils;
+import org.bukkit.DyeColor;
+import org.bukkit.entity.Entity;
+import org.bukkit.entity.Wolf;
+
+public class WolfSerializer extends EntitySerializer {
+    protected String color = null;
+
+    @Override
+    protected void serializer(Entity entity) {
+        color = ((Wolf) entity).getCollarColor().name().toLowerCase();
+    }
+
+    @Override
+    protected void deserializer(Entity entity) {
+        ((Wolf) entity).setCollarColor(MiscUtils.getEnum(color, DyeColor.RED));
+    }
+
+    @Override
+    protected void niceName(StringBuilder sb, int start) {
+        if (color != null) {
+            sb.insert(start, MiscUtils.niceName(color)).insert(start + color.length(), ' ');
+        }
+    }
+}
